text/md_to_docx/requirements.txt
python-docx==0.8.11
Markdown==3.4.3
beautifulsoup4==4.12.2
python-dateutil==2.8.2